backend/requirements.txt
boto3==1.20.53
botocore==1.23.53
Django==4.0.2
django-cors-headers==3.11.0
django-csp==3.7
django-imagekit==4.1.0
django-oauth-toolkit==1.7.0
django-rest-framework==0.1.0
django-storages==1.12.3
django-user-agents==0.4.0
django-user-visit==0.5.1
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==8.4.0
python-decouple==3.4
social-auth-app-django==5.0.0
social-auth-core==4.2.0
gunicorn